Moving editor outside of helper as recommended by Jpople
