+++ clients/admin-ui/src/features/connector-templates/SaaSVersionModal.tsx
@@ -1,13 +1,35 @@
 import { Button, Flex, Modal, Spin, Tabs, Text } from "fidesui";
 import React from "react";
 
+import ClipboardButton from "~/features/common/ClipboardButton";
 import { MODAL_SIZE } from "~/features/common/modals/modal-sizes";
-import { EditorWithCopy } from "~/features/common/yaml/helpers";
+import { Editor } from "~/features/common/yaml/helpers";
 
 import {
   useGetConnectorTemplateVersionConfigQuery,
   useGetConnectorTemplateVersionDatasetQuery,
 } from "./connector-template.slice";
+
+const EditorWithCopy = ({ yaml }: { yaml: string }) => (
+  <div style={{ position: "relative" }}>
+    <Editor
+      defaultLanguage="yaml"
+      value={yaml}
+      height="60vh"
+      options={{
+        readOnly: true,
+        minimap: { enabled: false },
+        fontSize: 13,
+        fontFamily: "Menlo",
+        scrollBeyondLastLine: false,
+      }}
+      theme="light"
+    />
+    <div style={{ position: "absolute", right: 8, top: 8, zIndex: 10 }}>
+      <ClipboardButton copyText={yaml} size="small" />
+    </div>
+  </div>
+);
 
 interface SaaSVersionContentProps {
   connectorType: string;
@@ -40,14 +62,14 @@ const SaaSVersionContent = ({
     }
     if (configError) {
       return (
-        <Text type="danger" className="text-sm">
+        <Text type="danger">
           Could not load version config.
         </Text>
       );
     }
     if (!configYaml) {
       return (
-        <Text type="secondary" className="text-sm">
+        <Text type="secondary">
           No config available for this version.
         </Text>
       );
@@ -65,14 +87,14 @@ const SaaSVersionContent = ({
     }
     if (datasetError) {
       return (
-        <Text type="danger" className="text-sm">
+        <Text type="danger">
           Could not load dataset for this version.
         </Text>
       );
     }
     if (!datasetYaml) {
       return (
-        <Text type="secondary" className="text-sm">
+        <Text type="secondary">
           No dataset available for this version.
         </Text>
       );
